API: Cstruct.ts have a built-in notion of 'offset' and 'length'
Rather than pass in additional 'offset' and 'length' parameters,
the client can use the 'Cstruct.sub' function to create a "view"
over the same underlying buffer.

Future versions of Cstruct (> 1.0.1) have extensive bounds-checking
built-in, which we get for free if we encourage the use of
'Cstruct.sub'. For now we keep our additional bounds-checks so
that we're safe for all versions of Cstruct.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>